web/tvirus/requirements.txt
beautifulsoup4==4.9.1
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
fastapi==0.61.1
h11==0.9.0
idna==2.10
lxml==4.5.2
pydantic==1.6.1
python-dateutil==2.8.1
python-multipart==0.0.5
redis==3.5.3
requests==2.24.0
six==1.15.0
soupsieve==2.0.1
SQLAlchemy==1.3.19
starlette==0.13.6
urllib3==1.25.10
uvicorn==0.11.8
websockets==8.1
